pyproject.toml
[project]
name = "turbo"
version = "0.1.0"
description = "JARVIS Orchestrator - Multi-engine AI agent powered by Claude Agent SDK"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.35",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "ccxt>=4.0.0",
    "textual>=1.0.0",
    "pystray>=0.19.0",
    "pillow>=10.0.0",
    "keyboard>=0.13.5",
    "numpy>=2.0.0",
    "sounddevice>=0.5.0",
    "fastapi>=0.129.0",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
    "openwakeword>=0.6.0",
    "edge-tts>=6.1.0",
    "faster-whisper>=1.1.0",
    "onnxruntime>=1.24.2",
    "scipy>=1.15.3",
    "datasets>=4.6.1",
    "pytest-asyncio>=1.3.0",
]

[project.scripts]
turbo = "main:main_sync"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"
addopts = "--ignore=tests/test_final.py --ignore=tests/test_pipelines_batch2.py --ignore=tests/test_pipelines_batch3.py --ignore=tests/test_pipelines_batch4.py --ignore=tests/test_pipelines_batch5.py --ignore=tests/test_pipelines_batch6.py --ignore=tests/test_pipelines_batch7.py --ignore=tests/test_pipelines_live.py --ignore=tests/test_domino_live.py --ignore=tests/test_domino_executor.py --ignore=tests/test_pattern_agents.py --ignore=tests/test_quick.py --ignore=tests/test_dashboard.py --ignore=tests/test_display_manager.py"
filterwarnings = [
    "ignore::RuntimeWarning",
    "ignore::DeprecationWarning",
]
